+++ R/write_codemeta.R
@@ -7,6 +7,7 @@
 #' `create_codemeta()`.
 #'
 #' @includeRmd man/rmdhunks/whybother.Rmd
+#' @includeRmd man/rmdhunks/uptodate.Rmd
 #'
 #' @param pkg package path to package root, or package name, or description file
 #'   (character), or a codemeta object (list)
@@ -18,9 +19,7 @@
 #'   \code{base::files.ize()}. Files in \code{.Rbuildignore} are ignored.
 #' @param force_update Update guessed fields even if they are defined in an
 #'   existing codemeta.json file
-#' @param use_git_hook Whether to create a pre-commit hook requiring
-#'   codemeta.json to be updated when DESCRIPTION is changed. Default is \code{FALSE} to avoid
-#'   an unwanted alteration of the user's git environment.
+#' @param use_git_hook Deprecated argument.
 #' @param verbose Whether to print messages indicating opinions e.g. when
 #'   DESCRIPTION has no URL, see \code{\link{give_opinions}}; and indicating
 #'    progress of internet downloads.
@@ -31,11 +30,6 @@
 #'   existing data in that block. In this case, the package root directory
 #'   should be the current working directory.
 #'
-#' When creating and writing a codemeta.json for the first time, the function
-#' adds "codemeta.json" to .Rbuildignore and, if the project uses git, adds a
-#' pre-commit hook ensuring that if DESCRIPTION changes, the codemeta.json will
-#' be updated as well unless the DESCRIPTION change is committed with 'git
-#' commit --no-verify'.
 #' @return writes out the codemeta.json file
 #' @export
 #'
@@ -45,9 +39,13 @@
 #' }
 write_codemeta <- function(
   pkg = ".", path = "codemeta.json", root = ".", id = NULL, use_filesize = FALSE,
-  force_update = getOption("codemeta_force_update", TRUE), use_git_hook = FALSE,
+  force_update = getOption("codemeta_force_update", TRUE), use_git_hook = NULL,
   verbose = TRUE, ...
 ) {
+
+  if (!missing(use_git_hook)) {
+    warning("The use_git_hook argument is deprecated and ignored.")
+  }
 
   codemeta_json <- "codemeta.json"
 
@@ -59,23 +57,7 @@ write_codemeta <- function(
   if (in_package && path == codemeta_json) {
 
     usethis::use_build_ignore(codemeta_json)
-
-    # Add the git pre-commit hook
-    # https://github.com/r-lib/usethis/blob/master/inst/templates/
-    #   readme-rmd-pre-commit.sh#L1
-    # this is GPL-3 code
-    no_codemeta_json <- ! file.exists(file.path(pkg, codemeta_json))
-
-    if (uses_git() && use_git_hook && no_codemeta_json) {
-
-      message(get_message("adding_hook")) # nolint
-
-      usethis::use_git_hook("pre-commit", render_template(
-        "description-codemetajson-pre-commit.sh"
-      ))
-    }
   }
-
   # Create or update codemeta and save to disk
   create_codemeta(pkg = pkg, root = root, use_filesize = use_filesize) %>%
     jsonlite::write_json(path, pretty = TRUE, auto_unbox = TRUE, ...)